Support incremental nodejs_install

+++ tools/nodejs/install.sh
@@ -5,6 +5,10 @@ export RUNFILES_DIR="$0.runfiles"
 
 cd "$BUILD_WORKSPACE_DIRECTORY"
 
+if [ -f node_modules/1.fix ]; then
+  exit
+fi
+
 # fix Angular language service
 rm -fr node_modules/@angular/core
 mv node_modules/.content/npm_angular_core_16.0.0-dc3fc578/files node_modules/@angular/core
@@ -14,3 +18,5 @@ mkdir -p node_modules/@angular/core/node_modules
 ln -s ../../../.content/npm_rxjs_7.4.0/files node_modules/@angular/core/node_modules/rxjs
 ln -s ../../../.content/npm_tslib_2.3.1/files node_modules/@angular/core/node_modules/tslib
 ln -s ../../../.content/npm_zone.js_0.11.4/files node_modules/@angular/core/node_modules/zone.js
+
+> node_modules/1.fix
